Fix compile-time bugs for global variables
Fix bugs that prevents the code from compiling. The bugs were not
noticed before because the code was tested with the OAI equivalent
of 'make' rather than 'make clean' and as such the changes to the
header files was not taken into account.

The code now compiles correctly, with all the global variables
used as expected.